Added some new terrain testing things and world chunks. Added primitive restart option and Getters to index buffer to get the restart value.

+++ RTS_Engine/Engine/World/World.h
@@ -16,7 +16,7 @@ namespace Bennett
 
 		bool m_IsLoaded;
 		std::unordered_map<std::string, Entity*> m_Entities;
-		std::unordered_map<glm::vec2, WorldChunk*> m_SpatialGrid;
+		std::unordered_map<glm::ivec2, WorldChunk*> m_SpatialGrid;
 
 	protected:
 		void DeleteAllEntities();
